% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motifFootprint.R
\name{addInsertionBias}
\alias{addInsertionBias}
\title{Add InsertionBias to MOCHA object}
\usage{
addInsertionBias(SampleTileObj, numCores = 1, verbose = TRUE)
}
\arguments{
\item{SampleTileObj}{A Sample-tile object from MOCHA's getSampleTileMatrix()}

\item{numCores}{Number of cores to parallelize oer}

\item{verbose}{A boolean to determin}
}
\value{
A Sample-Tile object with an insertion bias matrix within the metadata list
}
\description{
Generated and Add InsertionBias to MOCHA object
}
\keyword{downstream}
