% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{asMochaSTM}
\alias{asMochaSTM}
\title{Coerce to MochaSampleTileMatrix}
\usage{
asMochaSTM(x)
}
\arguments{
\item{x}{A RangedSummarizedExperiment from \code{getSampleTileMatrix}.}
}
\value{
A \code{MochaSampleTileMatrix} object.
}
\description{
Coerce to MochaSampleTileMatrix
}
\keyword{utils}
