% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{asMochaTileResults}
\alias{asMochaTileResults}
\title{Coerce to MochaTileResults}
\usage{
asMochaTileResults(x)
}
\arguments{
\item{x}{A MultiAssayExperiment from \code{callOpenTiles}.}
}
\value{
A \code{MochaTileResults} object.
}
\description{
Coerce to MochaTileResults
}
\keyword{utils}
