% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionalityReduction.R
\name{bulkLSI}
\alias{bulkLSI}
\title{\code{bulkLSI}}
\usage{
bulkLSI(SampleTileObj, cellType = "All", componentNumber = 30, verbose = FALSE)
}
\arguments{
\item{SampleTileObj}{The SummarizedExperiment object output from getSampleTileMatrix}

\item{componentNumber}{integer. Number of components to include in LSI. This must be strictly less than
the number of samples times the number of cellTypes in your SampleTileObj.}

\item{verbose}{Set TRUE to display additional messages. Default is FALSE.}

\item{cellPopulations}{vector of strings. Cell subsets for which to call
peaks. This list of group names must be identical to names that appear in
the SampleTileObj.  Optional, if cellPopulations='ALL', then peak
calling is done on all cell populations. Default is 'ALL'.}
}
\value{
LSIObj a SummarizedExperiment containing PC components from the LSI and metadata from the SampleTileObj
}
\description{
\code{bulkLSI} generates LSI (Latent semantic indexing)
}
\examples{
\dontrun{
LSIObj <- MOCHA::bulkLSI(SampleTileObj, cellType = 'CD16_Mono')
}
}
