% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimensionalityReduction.R
\name{bulkUMAP}
\alias{bulkUMAP}
\title{\code{bulkUMAP}}
\usage{
bulkUMAP(LSIObj, components = c(1:30), n_neighbors = 15)
}
\arguments{
\item{LSIObj}{The SummarizedExperiment object output from bulkLSI.}

\item{components}{A vector of integers. Number of components to include in LSI (1:30 typically).}

\item{n_neighbors}{See  \link[uwot]{umap}. The size of local neighborhood (in terms of number of
neighboring sample points) used for manifold approximation. Default is 15.}
}
\value{
fullUMAP data.frame of UMAP values with metadata attached.
}
\description{
\code{bulkUMAP} generates UMAP from pseudobulk LSIObj object, and merges in metadata.
}
\examples{
\dontrun{
UMAPvalues <- MOCHA::bulkUMAP(LSIObj)
}
}
