% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_intensities.R
\name{calculate_intensities}
\alias{calculate_intensities}
\title{\code{calculate_intensities}}
\usage{
calculate_intensities(fragMat, candidatePeaks, FALSE)
}
\arguments{
\item{fragMat}{a genomic ranges object containing the fragments}

\item{candidatePeaks}{a genomic ranges object indicating where to call peaks. This could be a pre-selected peak set or a dynamic bins approach to searching the whole genome.}

\item{totalFrags}{a numeric normalizing value, indicating the total # of fragments for a cell pop'n in a sample}
}
\value{
a data.table, countsByBin, that returns the two intensity parameters required
to calculate the probability of a (+) peak
}
\description{
\code{calculate_intensities} is an R helper function, part of the single-cell peak calling
algorithm scMACS by (XXX et al, 2021) that calculates the design matrix, X,
used to make peak calls.
}
\details{
The technical details of the algorithm are found in XX.
}
\references{
XX
}
