% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_local_noise.R
\name{calculate_local_noise}
\alias{calculate_local_noise}
\title{\code{calculate_local_noise}}
\usage{
calculate_local_noise(lambda1, width = 20)
}
\arguments{
\item{lambda1}{a vector, intensity (reads per cell) from which to calculate the 
background noise in a neighborhood around a given region}

\item{width}{integer, an positive integer indicating how many bins around region j
       to calculate background noise. For example, for 500-bp regions, width=20 will 
.       calculate noise-level in a 10kbp neighborhood.}
}
\value{
local_noise_estimates a vector length(lambda1) containing noise estimates 
        around each value of lambda1. For windows at the boundaries with < Width/2 
        neighbors either to the right or left, a 1-sided background noise estimate
        is provided using a one-sided search.
}
\description{
\code{calculate_local_noise} is an R helper function, part of the single-cell peak calling algorithm scMACS by (XXX et al, 2021) that determines the "background noise" located 
around each candidate peak region.
}
\references{
XX
}
