% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callOpenTiles.R
\name{callOpenTiles}
\alias{callOpenTiles}
\title{\code{callOpenTiles}}
\usage{
callOpenTiles(
  ArchRProj,
  cellPopLabel,
  cellPopulations = "ALL",
  TxDb = TxDb.Hsapiens.UCSC.hg38.refGene,
  Org = org.Hs.eg.db,
  outDir = NULL,
  fast = FALSE,
  numCores = 30,
  force = FALSE
)
}
\arguments{
\item{ArchRProj}{an ArchR Project}

\item{cellPopLabel}{string indicating which column in the ArchRProject
metadata contains the cell population label.}

\item{cellPopulations}{vector of strings. Cell subsets for which to call
peaks. This list of group names must be identical to names that appear in
the ArchRProject metadata.  Optional, if cellPopulations='ALL', then peak
calling is done on all cell populations in the ArchR project metadata.
Default is 'ALL'.}

\item{TxDb}{is an AnnotationDbi object with transcript info for the organism. UCSC Hg38 Refgene is the default.}

\item{outDir}{is a string describing the output directory for coverage files per sample/celltype. 
Must be a complete directory string. Default is NULL, in which case it'll pull out a directory from ArchR 
and make a new fold named MOCHA for saving files.}

\item{numCores}{integer. Number of cores to parallelize peak-calling across
multiple cell populations.}
}
\value{
tileResults A MultiAssayExperiment object containing ranged data
  for each tile
}
\description{
\code{callOpenTiles} is the main peak-calling function in scMACS
  that serves as a wrapper function to call peaks provided a set of fragment
  files and an ArchR Project for meta-data purposes
}
\details{
The technical details of the algorithm are found in XX.
}
\references{
XX
}
