% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeaks_population.R
\name{callPeaks_by_population}
\alias{callPeaks_by_population}
\title{\code{callPeaks}}
\usage{
callPeaks_by_population(
  blackList,
  returnAllPeaks = FALSE,
  numCores = 10,
  totalFrags,
  fragsList,
  StudypreFactor
)
}
\arguments{
\item{blackList}{A GRanges object containing a blacklist of regions to exclude}

\item{returnAllPeaks}{boolean. Indicates whether scMACS should return object containing all genomic regions or just the positive (+) called peaks. Default to the latter, only positive peaks.}

\item{numCores}{integer. Number of cores to parallelize peak-calling across
multiple cell populations}

\item{totalFrags}{# of fragments in that sample for that cell population}

\item{StudypreFactor}{ratio of average signal between new study and training set}

\item{cellColData}{The cell-level metadata from the original ArchR Project}

\item{cellPopulation}{String/character vector. Cell subset for which to call peaks.}

\item{cellPopLabel}{string indicating which column in the ArchRProject metadata contains 
the cell population label.}
}
\value{
scMACs_PeakList an list containing peak calls for each cell population passed on in the 
        cell subsets argument. Each peak call is returned as as Genomic Ranges object.
}
\description{
\code{callPeaks} is the main peak-calling function in scMACS
             that serves as a wrapper function to call peaks provided a set 
             of fragment files and cell metadata
}
\details{
The technical details of the algorithm are found in XX.
}
\references{
XX
}
