% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeaks_sample.R
\name{callPeaks_by_sample}
\alias{callPeaks_by_sample}
\title{\code{callPeaks_by_sample}}
\usage{
callPeaks_by_sample(
  ArchRProj,
  cellPopLabel,
  cellPopulations = "ALL",
  returnAllPeaks = TRUE,
  numCores = 30,
  returnFrags = F
)
}
\arguments{
\item{ArchRProj}{an ArchR Project}

\item{cellPopLabel}{string indicating which column in the ArchRProject metadata contains 
the cell population label.}

\item{cellPopulations}{vector of strings. Cell subsets for which to call peaks. This list of group names must be identical to names that appear in the ArchRProject metadata.  Optional, if cellPopulations='ALL', then peak calling is done on all cell populations in the ArchR project metadata. Default is 'ALL'.}

\item{returnAllPeaks}{boolean. Indicates whether scMACS should return object containing all genomic regions or just the positive (+) called peaks. Default to the latter, only positive peaks.}

\item{numCores}{integer. Number of cores to parallelize peak-calling across
multiple cell populations}

\item{returnFrags}{boolean. Determines if fragment files should be returned}
}
\value{

}
\description{
\code{callPeaks_by_sample} is the main peak-calling function in scMACS
             that serves as a wrapper function to call peaks provided a set 
             of fragment files and an ArchR Project for meta-data purposes
}
\details{
The technical details of the algorithm are found in XX.
}
\references{
XX
}
