% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocha_subclass_methods.R
\name{cellTypes}
\alias{cellTypes}
\title{Get cell population names from a MOCHA object}
\usage{
cellTypes(x)
}
\arguments{
\item{x}{A \code{MochaTileResults} or \code{MochaSampleTileMatrix} object.}
}
\value{
Character vector of cell population names.
}
\description{
Get cell population names from a MOCHA object
}
\keyword{utils}
