% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctMOCHA.R
\name{correctGenome}
\alias{correctGenome}
\title{\code{correctGenome}}
\usage{
correctGenome(TSAM, genome)
}
\arguments{
\item{TSAM}{RangedSummarizedExperiment object, containing the TSAM from getSampleTileMatrix, with a motif set added via addMotifSet()}

\item{genome}{genome to correct with.}
}
\value{
chromVAR object
}
\description{
\code{correctGenome} Runs ChromVAR on a MOCHA Tile-Sample Accessibility Matrix object
}
\details{
This is a wrapper for basic SummarizedExperiment-based subsetting.
}
\keyword{utils}
