% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Region.R
\name{countdf_to_region}
\alias{countdf_to_region}
\title{CountDF to Region GRanges}
\usage{
countdf_to_region(countdf)
}
\arguments{
\item{countdf}{A dataframe that comes from `getbpCounts()` or `getbpInserts()`. Expected columns "chr" and "Locus"}
}
\value{
A granges object for the region defined in countdf
}
\description{
Gets the range covered by counts in a count df and converts to granges
Used in `plotRegion()`
}
