% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingUtils.R
\name{counts_plot_motif_overlay}
\alias{counts_plot_motif_overlay}
\title{Plot Motifs on Counts Plot}
\usage{
counts_plot_motif_overlay(
  p1,
  countdf,
  motifsList,
  motif_y_space_factor = 4,
  motif_stagger_labels_y = FALSE,
  motif_weights = NULL,
  motif_weight_name = "Motif Weight",
  motif_weight_colors = c(darkblue = -10, gray = 0, darkred = 10),
  motif_lab_size = 1,
  motif_lab_alpha = 0.25,
  motif_line_size = 0.75,
  motif_line_alpha = 0.25
)
}
\arguments{
\item{p1}{The output of `counts_plot_samples()`}

\item{countdf}{A dataframe that comes from `getbpCounts()` or `getbpInserts()`}

\item{motif_y_space_factor}{A factor for vertical spacing between motif labels. Default 4. Increase to make labels farther apart, decrease to make labels closer.}

\item{motif_stagger_labels_y}{= FALSE Logical value, default FALSE. If TRUE, will  stagger motif labels in adjacent columns in the vertical direction}

\item{motif_weights}{Optional numeric vector, default NULL. If provided will be used to color motif labels by the weighted values. Values must be uniquely named
with motif names, for example c(`KLF5`= 3.2, `STAT1 = 0.2`, `EOMES` = -1.4`). Weights can be anything relevant, for example if the peak/region is associated with
a specific group/sample then global motif enrichment results for that group: `-log10(FDR)*sign(change)`}

\item{motif_weight_name}{Character value, default "Motif Weight". Used to label the color legend.}

\item{motif_lab_size}{Numeric value, default 1. Size of motif labels.}

\item{motif_lab_alpha}{Numeric value, default 0.25. Alpha for motif labels.}

\item{motif_line_size}{Numeric value, default 1. Size of motif lines.}

\item{motif_line_alpha}{Numeric value, default 0.25. Alpha for motif lines.}

\item{ArchRProj}{An archR project containing motif annotations}

\item{motifSetName}{The name of the motif set in ArchRProj to use for annotation}

\item{weight_colors}{Named numeric vector. Names should be color values and breaks should be the corresponding values of motif_weights. Values outside the
highest and lowest value will appear as max or min defined color value.}
}
\value{
The input ggplot object with motif labels overlaid
}
\description{
Add motif annotation labels on an existing counts plot. If providing weights for motif
text colors, ensure the input counts plot is formatted accordingly (ie counts_color = "gray90")
for best visibility.
}
