% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingUtils.R
\name{counts_plot_samples}
\alias{counts_plot_samples}
\title{Plot normalized counts for each sample}
\usage{
counts_plot_samples(
  countdf,
  plotType = "area",
  base_size = 12,
  counts_color = NULL,
  range_label_size = 4,
  legend.position = NULL,
  facet_label_side = "top",
  counts_group_colors = NULL,
  counts_color_var = "Groups",
  theme_ls = .counts_plot_default_theme
)
}
\arguments{
\item{countdf}{A dataframe that comes from `getbpCounts()` or `getbpInserts()`}

\item{plotType}{Options include 'overlaid','area', 'line', or 'RidgePlot'. default is 'area', which will plot a seperate track for each group.
Setting plotType to 'overlaid' will overlay count plot histograms across samples, instead of faceting out separately.
Setting plotType to 'RidgePlot' will generate a ridgeplot across all groups.}

\item{base_size}{Numeric, default 12. Global plot base text size parameter}

\item{counts_color}{Optional color palette. A named vector of color values where names are unique values in the `color_var` column}

\item{range_label_size}{Numeric value, default 4. Text size for the y-axis range label}

\item{legend.position}{Any acceptable `legend.position` argument to theme(). Default NULL will place legend for overlaid plots at (0.8,0.8),
or to the "right" for faceted plots.}

\item{facet_label_side}{Direction character value, default "top". Can also be "right", "left", or "bottom". Position of facet label.}

\item{counts_group_colors}{Optional named color vector. Values as colors, names are levels of `counts_color_var`. If provided, will color the plots specifically
using `scale_color_manual()`}

\item{counts_color_var}{Character value, default "Groups". Column name from countdf to use to color counts plots.
Only used if counts_group_colors provided}

\item{theme_ls}{A list of named theme arguments passed to theme(), defaults to `.counts_plot_default_theme`. For example, `list(axis.ticks = ggplot2::element_blank())`}
}
\value{
A ggplot object of count histograms by sample.
}
\description{
Used in `plotRegion()` for the counts tracks
}
