% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_sample_peak_matrix.R
\name{create_peak_sampleMatrix}
\alias{create_peak_sampleMatrix}
\title{\code{create_peak_sampleMatrix}}
\usage{
create_peak_sampleMatrix(sample_specific_peaks, union_peaks)
}
\arguments{
\item{sample_specific_peaks}{output of calls_peaks_by_sample (peaks only)}

\item{union_peaks:}{a vector containing the tileIDs to subset the peak matrix}
}
\value{
sample_peak_matrix a sample X peak matrix containing observed 
        measurements for each sample at each peak.
}
\description{
\code{create_peak_sampleMatrix} is a function that can 
             transform a set of sample-specific peak calls into 
             peak X sample matrix containing lambda1 measurements for 
             each sample.
}
\details{
The technical details of the algorithm are found in XX.
}
\references{
XX
}
