% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_dynamic_range.R
\name{determine_dynamic_range}
\alias{determine_dynamic_range}
\title{\code{determine_dynamic_range}}
\usage{
determine_dynamic_range(AllFragmentsList, cellSubsetArchR, binSize, doBin)
}
\arguments{
\item{AllFragmentsList:}{List of fragments by arrow file}

\item{GeneralWindowSize:}{Window size for sliding window generated over longer fragments.}

\item{WindowSizeRange:}{The sliding window function will generate a smaller window at the end of a longer fragment,}

\item{coreNum}{an integer indicating the number of cores to use}

\item{doBins}{is a boolean variable. When true, then it will into windows according to GeneralWindowSize and WindowSizeRange.}
}
\value{
a data.table, countsByBin, that returns the two intensity parameters required
to calculate the probability of a (+) peak
}
\description{
\code{determine_dynamic_range} is an R helper function, part of the single-cell peak calling
algorithm scMACS by (XXX et al, 2021) that determines which genomic regions, or bins,
will be used for de-novo peak calling. The function "determine_dynamic_range" is a wrapper function
around the dynamic_bins function
}
\details{
The technical details of the algorithm are found in XX.
}
\references{
XX
}
