% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_accessibility.R
\name{differential_accessibility}
\alias{differential_accessibility}
\title{\code{differential_accessibility}}
\usage{
differential_accessibility(groupA, groupB)
}
\arguments{
\item{groupA:}{list of accessible peaks for each sample in group A}

\item{groupB:}{list of accessible peaks for each sample in group B}
}
\value{
a genomic ranges object that 
to calculate the probability of a (+) peak
}
\description{
\code{differential_accessibility} allows you to determine whether regions of chromatin are 
             differentially accessible betwen groups by conducting a test
}
\details{
The technical details of the algorithm are found in XX.
}
\references{
XX
}
