% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingUtils.R
\docType{data}
\name{.counts_plot_default_theme}
\alias{.counts_plot_default_theme}
\title{Default ggplot theme for counts plot}
\format{
An object of class \code{list} of length 10.
}
\usage{
.counts_plot_default_theme
}
\description{
Default ggplot theme for counts plot
}
\keyword{datasets}
