% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingUtils.R
\docType{data}
\name{.gene_plot_theme}
\alias{.gene_plot_theme}
\title{Common theme for gene plots}
\format{
An object of class \code{list} of length 5.
}
\usage{
.gene_plot_theme
}
\description{
Common theme for gene plots
}
\keyword{datasets}
