% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_sample_celltype_count_tables}
\alias{.get_sample_celltype_count_tables}
\title{Extract Sample-celltype specific metadata}
\usage{
.get_sample_celltype_count_tables(object)
}
\arguments{
\item{object}{tileResults object from callOpenTiles or SummarizedExperiment
from getSampleTileMatrix}
}
\value{
a SummarizedExperiment where each assay is a different type of
  metadata.
}
\description{
\code{getSampleCellTypeMetadata} Extract Sample-celltype
  specific metadata like fragment and cell counts
}
\keyword{utils}
