% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_bins.R
\name{dynamic_bins}
\alias{dynamic_bins}
\title{\code{dynamic_bins}}
\usage{
dynamic_bins(
  AllFragmentsList,
  GeneralWindowSize,
  WindowSizeRange,
  coreNum,
  doBin
)
}
\arguments{
\item{coreNum}{an integer indicating the number of cores to use}

\item{AllFragmentsList:}{List of fragments by arrow file}

\item{GeneralWindowSize:}{Window size for sliding window generated over longer fragments.}

\item{WindowSizeRange:}{The sliding window function will generate a smaller window at the end of a longer fragment,}

\item{doBins}{is a boolean variable. When true, then it will into windows according to GeneralWindowSize and WindowSizeRange.}
}
\value{
a data.table, countsByBin, that returns the two intensity parameters required
to calculate the probability of a (+) peak
}
\description{
\code{dynamic_bins} is an R helper function, part of the single-cell peak calling
algorithm scMACS by (XXX et al, 2021) that determines which genomic regions, or bins,
will be used for de-novo peak calling. The function "dynamic_bins" generates bins for peak calling based on the actual fragments present within a sample.
}
\details{
The technical details of the algorithm are found in XX.
}
\references{
XX
}
