% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_accessibility.R
\name{estimate_differential_accessibility}
\alias{estimate_differential_accessibility}
\title{\code{differential_accessibility}}
\usage{
estimate_differential_accessibility(
  tile_values,
  group,
  providePermutation = FALSE
)
}
\arguments{
\item{tmp_mat:}{sample-peak matrix}

\item{tileID:}{chromosome region to analyze}

\item{group:}{vector of group indices (0,1)}

\item{savePlotToFile:}{boolean: do you want to save this file. Defaults to current directory}
}
\value{
a table indicating the results for differential accessibility
        which includes the following pieces of informations
        - Peak= Peak ID
        - P_value= P-value from two-part wilcoxon test
        - TestStatistic= statistic from two-part wilcoxon test
        - HL_EffectSize=hodges_lehmann effect size,
        - Case_mu= Median of nonzero values in case condition,
        - Case_rho=proportion of zeros in case condition ,
        - Control_mu= Median of nonzero values in control condition
        - Control_rho==proportion of zeros in control condition ,
}
\description{
\code{differential_accessibility} allows you to determine whether regions of chromatin are 
             differentially accessible betwen groups by conducting a test
}
\details{
The technical details of the algorithm are found in XX.
}
\references{
XX
}
