% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{exampleBlackList}
\alias{exampleBlackList}
\title{exampleBlackList}
\format{
A GRanges object with 210 ranges and 2 metadata columns
}
\usage{
exampleBlackList
}
\description{
Example input of a blackList extracted from the PBMC_Small dataset
consisting of 2k cells and spanning chr1 and 2 (~2-300MB).
The data is publicly available with the ArchR package at
<https://www.archrproject.com/reference/getTestProject.html>
}
\keyword{datasets}
