% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{exampleCellColData}
\alias{exampleCellColData}
\title{exampleCellColData}
\format{
A DataFrame with 2217 rows and 3 columns
}
\usage{
exampleCellColData
}
\description{
Example input of cellColData extracted from the PBMC_Small dataset
consisting of 2k cells and spanning chr1 and 2 (~2-300MB).
The data is publicly available with the ArchR package at
<https://www.archrproject.com/reference/getTestProject.html>
}
\keyword{datasets}
