% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{exampleFragments}
\alias{exampleFragments}
\title{exampleFragments}
\format{
A list of 2 GRanges objects
}
\usage{
exampleFragments
}
\description{
Example input of ATAC fragments extracted from the PBMC_Small dataset
consisting of 2k cells and spanning chr1 and 2 (~2-300MB).
This subset consists of two cell populations: Clusters C2 and C5.
The data is publicly available with the ArchR package at
<https://www.archrproject.com/reference/getTestProject.html>
}
\keyword{datasets}
