% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{exampleGenome}
\alias{exampleGenome}
\title{exampleGenome}
\format{
A BSgenome object containing annotations for the hg19 genome
}
\usage{
exampleGenome
}
\description{
Example input of a genome extracted from the PBMC_Small dataset 
consisting of 2k cells and spanning chr1 and 2 (~2-300MB).
The data is publicly available with the ArchR package at
<https://www.archrproject.com/reference/getTestProject.html>
}
\keyword{datasets}
