% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{exampleTileResults}
\alias{exampleTileResults}
\title{exampleTileResults}
\format{
A MultiAssayExperiment object of 2 experiments
}
\usage{
exampleTileResults
}
\description{
Example output of MOCHA::callOpenTiles using exampleFragments.
Contains results on two cell populations: Clusters C2 and C5
}
\keyword{datasets}
