% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportInsertions.R
\name{exportLocalFootprints}
\alias{exportLocalFootprints}
\title{Export insertion counts to per-sample BigWig files after applying
  a rolling sum and rolling median smoothing filter.}
\usage{
exportLocalFootprints(
  SampleTileObj,
  cellPopulation,
  outDir = NULL,
  windowSize = 10,
  groupColumn = NULL,
  subGroups = NULL,
  sampleSpecific = FALSE,
  normTn5 = TRUE,
  force = FALSE,
  slow = FALSE,
  verbose = FALSE,
  numCores = 5
)
}
\arguments{
\item{SampleTileObj}{A MultiAssayExperiment or RangedSummarizedExperiment
from MOCHA}

\item{cellPopulation}{A string denoting the cell population of interest}

\item{outDir}{Directory to write output bigwig files. Default is NULL, where
the directory in `SampleTileObj@metadata$Directory` will be used.}

\item{windowSize}{Window size for rolling sum & median over basepairs. Default is 10.}

\item{groupColumn}{A string, corresponding to a metadata column within the SampleTileObj, that describes the groups by which you want to summarize motif footprints. If sampleSpecific = FALSE, then the function will average insertions across samples within each group.}

\item{subGroups}{A list of subgroups, if you want to only look at specific groups within the groupColumn.}

\item{sampleSpecific}{A boolean for whether to generate average motif footprints within each group, or to return a data.frame for all samples.}

\item{normTn5}{A boolean for whether to normalize by Tn5 insertion bias.}

\item{force}{Set TRUE to overwrite existing files. Default is FALSE.}

\item{slow}{Set TRUE to bypass optimisations and compute smoothing filter
directly on the whole genome. May run slower and consume more RAM. Default
is FALSE.}

\item{verbose}{Set TRUE to display additional messages. Default is FALSE.}

\item{numCores}{Number of cores to parallelize over}
}
\value{
outPaths List of paths of exported insertion files
}
\description{
\code{exportLocalFootprints} Takes a SampleTileMatrix with
  linked insertion files, corrects Tn5 insertion bias, and applies a smoothing filter (a rolling sum then
  rolling median) to the insertions. This generates a local footprint track for visually identifying evidence of potential binding sites. 
  These files are then written to bigwig format.
}
\examples{
\dontrun{
# Depends on and manipulates files on filesystem
outPath <- MOCHA::exportSmoothedInsertions(
  SampleTileObj,
  cellPopulation = "CD4 Naive", sumWidth = 10, medianWidth = 11, verbose = FALSE
)
}

}
\keyword{exporting}
