% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlePopulationConsensusTiles.R
\name{extractErrorFromConsensusTiles}
\alias{extractErrorFromConsensusTiles}
\title{\code{extractErrorFromConsensusTiles}}
\usage{
extractErrorFromConsensusTiles(x)
}
\arguments{
\item{x}{a specific output from consensus tiles.}
}
\description{
\code{extractErrorFromConsensusTiles} is an R helper function, part of
  the single-cell peak calling.
}
\keyword{internal}
