% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCoAccessibleLinks.R
\name{filterCoAccessibleLinks}
\alias{filterCoAccessibleLinks}
\title{\code{filterCoAccessibleLinks}}
\usage{
filterCoAccessibleLinks(TileCorr, threshold = 0.5)
}
\arguments{
\item{TileCorr}{The correlation table output from getCoAccessibleLinks}

\item{threshold}{Keep}
}
\value{
FilteredTileCorr The filtered correlation table with chr,
  start, and end site of each link
}
\description{
\code{filterCoAccessibleLinks} will filter the output from
  getCoAccessibleLinks by a threshold, retaining links with a
  absolute correlation greater than the threshold.
  This function also adds the chr, start, and end site of each link to
  the output table.
}
\examples{
\dontrun{
# links is the output of MOCHA::getCoAccessibleLinks
MOCHA::filterCoAccessibleLinks(links, threshold = 0.5)
}

}
