% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{finalModelObject}
\alias{finalModelObject}
\title{finalModelObject}
\format{
A list of lists containing 2 items: "Loess" and "Linear" each with "Total" "Max" and "Intercept"
\describe{
\item{Loess}{LOESS model}
\item{Linear}{Linear model}
}
}
\usage{
finalModelObject
}
\description{
Trained MOCHA models - LOESS and linear regression
}
\keyword{datasets}
