% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_alpha_threshold.R
\name{find_alpha_threshold}
\alias{find_alpha_threshold}
\title{\code{plot_differential_region}}
\usage{
find_alpha_threshold(res_pvals, threshold)
}
\arguments{
\item{res_pvals:}{output from differential accessibility}

\item{threshold:}{fdr threshold to control for}
}
\value{
nominal_threshold: the largest nominal threshold 
        that controls for false discovery at the user-defined
        threshold.
}
\description{
\code{plot_differential_region} allows you to graph a particular
             region across two different conditions
}
