% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_peaks.R
\name{format_peaks}
\alias{format_peaks}
\title{\code{format_peaks.R}}
\usage{
format_peaks(peaks)
}
\arguments{
\item{peaks}{a peak matrix with a string containing genomic address,
'chr:XXX-XXX', where chromosome is delimited with colon,
               and start/end sites with a dash}
}
\value{
a granges object
}
\description{
\code{format_peaks.R} allows you to format data.tables into
             granges by creating the seqnames, start, end fields from a string
}
