% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_differential_accessible_regions.R
\name{get_differential_accessible_regions}
\alias{get_differential_accessible_regions}
\title{\code{get_differential_accessible_Regions}}
\usage{
get_differential_accessible_regions(
  sample_peak_matrix,
  metaFile,
  fdr_control = 0.2,
  nCores = 2
)
}
\description{
\code{get_differential_accessible_Regions} allows you to determine whether regions of chromatin are 
             differentially accessible betwen groups by conducting a test
}
