% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingUtils.R
\name{get_gene_body_model}
\alias{get_gene_body_model}
\title{Get Gene Body Model}
\usage{
get_gene_body_model(
  whichGene,
  regionGRanges,
  countdf,
  orgdb,
  db_id_col = "REFSEQ",
  verbose = TRUE
)
}
\arguments{
\item{whichGene}{Character value. Gene symbol for a single gene to plot.}

\item{regionGRanges}{Granges object for region of interest. Ie, an output of `countdf_to_region()`}

\item{countdf}{A dataframe that comes from `getbpCounts()` or `getbpInserts()`}

\item{orgdb}{An organism database containing the gene}

\item{db_id_col}{Character value. Column in `orgdb` containing the output id. Default "REFSEQ".}

\item{verbose}{Logical value, default TRUE.}
}
\value{
GRanges object for gene to plot
}
\description{
Get Gene Body model for specific gene in region ranges
}
\details{
Used in `plotRegion()` to preformat GRanges for selected genes
}
