% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Region.R
\name{get_motifs_in_region}
\alias{get_motifs_in_region}
\title{Get GRanges of motif annotations within a region}
\usage{
get_motifs_in_region(
  motifsList,
  countdf = NULL,
  regionString = NULL,
  numCores = 1,
  metaColumn = NULL,
  cellSubsets = "ALL"
)
}
\arguments{
\item{countdf}{A counts data frame object from getPop}

\item{ArchRProj}{The archr project containing motif annotations}

\item{motifSetName}{The name of the motif annotations in the ArchR project}
}
\description{
Helper to return the motif annotations within a specfic region outside of
the plotting function. Can supply either a counts df or a region string.
}
