% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Region.R
\name{get_promoter_range}
\alias{get_promoter_range}
\title{Get gene promoter range string}
\usage{
get_promoter_range(
  proj,
  gene,
  upstream = 50000,
  downstream = 50000,
  output_string = TRUE
)
}
\arguments{
\item{proj}{An ArchR Project}

\item{gene}{The name of a gene}

\item{upstream}{The number of basepairs to extend the range upstream of the gene promoter}

\item{downstream}{The number of basepairs to extend the range downstream of the gene promoter}

\item{output_string}{Logical value, default TRUE. If TRUE outputs a range string (ie 'chr1: 1700000-1710000') else
outputs a granges object}
}
\value{
A range string or granges object, depending on 'output_string' parameter.
}
\description{
For a given gene, return the gene position from an ArchR project
}
