% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getAnnotationDbFromInstalledPkgname}
\alias{getAnnotationDbFromInstalledPkgname}
\title{\code{getAnnotationDbFromInstalledPkgname} Loads and attaches an installed TxDb or
  OrgDb-class Annotation database package.}
\usage{
getAnnotationDbFromInstalledPkgname(dbName, type)
}
\arguments{
\item{dbName}{Exact name of installed annotation data package.}

\item{type}{Expected class of the annotation data package, must be
either "OrgDb" or "TxDb".}
}
\value{
the loaded Annotation database object.#' @noRd
}
\description{
See \link[BSgenome]{getBSgenome}
}
