% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCoAccessibility.R
\name{getBackGroundObj}
\alias{getBackGroundObj}
\title{\code{getBackGroundObj}}
\usage{
getBackGroundObj(STObj, NAtoZero = TRUE)
}
\arguments{
\item{STObj}{The SummarizedExperiment object output from getSampleTileMatrix containing your sample-tile matrices}

\item{NAToZero}{Set NA values in the sample-tile matrix to zero}
}
\value{
TileCorr A data.table correlation matrix
}
\description{
\code{getBackGroundObj} combines all celltypes in a SampleTileMatrix object into a SummarizedExperiment with one single matrix across all cell types and samples.
}
