% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCellPopMatrix.R
\name{getCellPopMatrix}
\alias{getCellPopMatrix}
\title{\code{getCellPopMatrix}}
\usage{
getCellPopMatrix(
  SampleTileObj,
  cellPopulation,
  dropSamples = TRUE,
  NAtoZero = TRUE
)
}
\arguments{
\item{SampleTileObj}{The output from getSampleTileMatrix, a
SummarizedExperiment of pseudobulk intensities across all tiles & cell
types.}

\item{cellPopulation}{The cell population you want to pull out.}

\item{dropSamples}{Boolean flag to determine whether to drop samples that
were too small for peak calling.}

\item{NAtoZero}{Boolean flag to determine whether to replace NAs with zero}
}
\value{
sampleTileMatrix a matrix of samples by called tiles for a given cell
  population.
}
\description{
\code{getCellPopMatrix} pulls out the SampleTileMatrix of tiles
  called in one given cell population.
}
