% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifEnrichment.R
\name{getCellTypeMotifs}
\alias{getCellTypeMotifs}
\title{Get motifset for a given cell type}
\usage{
getCellTypeMotifs(
  STM,
  cellPopulation,
  MotifSetName = "Motifs",
  asGRangesList = TRUE
)
}
\arguments{
\item{STM}{A Sample-Tile Object from MOCHA}

\item{cellPopulation}{The name of a cell type within MOCHA}

\item{MotifSetName}{The name of the total motif set, saved as metadata within the MOCHA object}

\item{asGRangesList}{Default is FALSE. If TRUE, then converts list of GRanges to GRangesList, which takes quite a while.}
}
\value{
A GRangesList containing the motifs that are present in open regions for a given cell type
}
\description{
Extracts an existing motif set and filters it down to only motifs found within regions open in that celltype
               A necessary step for accurate motif enrichment analysis.
}
\keyword{downstream}
