% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCellTypes}
\alias{getCellTypes}
\title{\code{getCellTypes} Extract cell type names from a Tile Results or Sample Tile object.}
\usage{
getCellTypes(object)
}
\arguments{
\item{object}{tileResults object from callOpenTiles or SummarizedExperiment from getSampleTileMatrix}
}
\value{
a vector of cell type names.
}
\description{
\code{getCellTypes} Returns a vector of cell names from a Tile Results or Sample Tile object.
}
