% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getCellTypeTiles}
\alias{getCellTypeTiles}
\title{\code{getCellTypeTiles} Extract the GRanges for a particular cell type}
\usage{
getCellTypeTiles(object, cellType)
}
\arguments{
\item{object}{A SampleTileObject.}

\item{cellType}{A string describing one cell type.}
}
\value{
a vector of cell type names.
}
\description{
\code{getCellTypeTiles} Returns a GRanges object of all tiles called for a certain cell type
}
