% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPopFrags.R
\name{getPopFrags}
\alias{getPopFrags}
\title{\code{getPopFrags}}
\usage{
getPopFrags(
  ArchRProj,
  metaColumn,
  cellSubsets = "ALL",
  region = NULL,
  numCores = 1,
  NormMethod = "NFrags",
  blackList = NULL,
  overlapList = 50
)
}
\arguments{
\item{ArchRProj}{an ArchR Project}

\item{metaColumn}{a string indicating which column to use to extract cells}

\item{cellSubsets}{vector of strings. Cell subsets for which to call peaks. Optional, if cellSubsets='ALL', then peak calling is done on all cell populations in the ArchR project metadata}

\item{region}{a chromosomal location. If NULL, it pull from the entire genome.}

\item{numCores}{integer number of cores to parallelize on}

\item{NormMethod}{= string indicating which normalization method to use. For scMACS modeling
we require the nFrags normalization}

\item{overlapList}{integer indicating acceptable overlap with blacklist regions in bp}

\item{blacklist}{a granges object indicating the blacklist region}
}
\value{
a list of fragments with a given normalization factor.
}
\description{
\code{getPopFrags} is an R helper function that extracts fragments  
             for a cell population from an archR Project
}
