% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reproducible_peaks.R
\name{getReproduciblePeaks}
\alias{getReproduciblePeaks}
\title{\code{get_reproducible_peaks}}
\usage{
getReproduciblePeaks(peakCallResults, cellPopulation, reproducibility = 0.2)
}
\description{
\code{get_reproducible_peaks} is an R helper function, part of the single-cell peak
}
