% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReproducibleTiles.R
\name{getReproducibleTiles}
\alias{getReproducibleTiles}
\title{\code{get_reproducible_peaks}}
\usage{
getReproducibleTiles(tileResults, cellPopulation, reproducibility = 0.2)
}
\description{
\code{get_reproducible_peaks} is an R helper function, part of
  the single-cell peak
}
