% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristicAnalysis.R
\name{getSequencingBias}
\alias{getSequencingBias}
\title{\code{getSequencingBias}}
\usage{
getSequencingBias(
  SampleTileObj,
  cellPopulations = "all",
  cellPopulation,
  groupColumn,
  foreground,
  background,
  verbose = TRUE
)
}
\arguments{
\item{SampleTileObj}{a SummarizedExperiment object generated by MOCHA}

\item{cellPopulations}{vector of strings. Cell subsets found in the TSAM, or the word 'All' if all should be used.}

\item{cellPopulation}{A string denoting the cell population of interest}

\item{groupColumn}{The column containing sample group labels}

\item{foreground}{The foreground group of samples for differential comparison}

\item{background}{The background group of samples for differential comparison}

\item{verbose}{Set TRUE to display additional messages. Default is FALSE.}
}
\value{
plot object
}
\description{
\code{getSequencingBias} takes the output of peak calling with
  callOpenTiles and creates sample-tile matrices containing the signal
  intensity at each tile.
}
