% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateFragments.R
\name{getSimulatedPeakSet}
\alias{getSimulatedPeakSet}
\title{simulate open and closed regions across the genome}
\usage{
getSimulatedPeakSet(
  peakNumber = 2e+05,
  peakSize = 500,
  peakPeriodicity = 10,
  Genome = BSgenome.Hsapiens.UCSC.hg38,
  allLocations = TRUE
)
}
\arguments{
\item{peakNumber}{Number of open tiles. Default is 200,000}

\item{peakSize}{Size of a given 'peak' or 'tile' in basepairs. Default is 500 bp.}

\item{peakPeriodicity}{Open regions tend to cluster. This metric measures the periodicity of that clusters in terms of number of tiles. Default is 10 tiles.
Resulting probability will be generated using a sin function, and all negative values will be set to zero, 
so only one cluster of open regions will occur within a given period.}

\item{Genome}{A BSgenome file. Default is BSgenome.Hsapiens.UCSC.hg38,}

\item{allLocations}{A boolean, which determines whether or not to return all potential tiles (with open regions flagged), or just open regions.}
}
\value{
A GenomicRanges object of regions
}
\description{
\code{getSimulatedPeakSet} Simulate
}
