% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GRangesToString}
\alias{GRangesToString}
\title{\code{GRangesToString} Converts a GRanges object to a string in the format 'chr1:100-200'}
\usage{
GRangesToString(GR_obj)
}
\arguments{
\item{GR_obj}{the GRanges object to convert to a string}
}
\value{
A string or list of strings in the format 'chr1:100-200' representing
 ranges in the input GRanges
}
\description{
\code{GRangesToString} Turns a GRanges Object into
 a list of strings in the format chr1:100-200
}
