% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runZIGLMM.R
\name{individualZIGLMM}
\alias{individualZIGLMM}
\title{\code{IndividualLMEM}}
\usage{
individualZIGLMM(x)
}
\arguments{
\item{refList.}{A list where the first index is a data.frame to use for modeling, and the second is the formula for modeling.}
}
\value{
A linear model
}
\description{
\code{IndividualLMEM} Runs linear modeling on data provided. Written for efficient parallelization.
}
