% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isMOCHAObject}
\alias{isMOCHAObject}
\title{isMOCHAObject}
\usage{
isMOCHAObject(Object, returnType = FALSE)
}
\arguments{
\item{Object}{The output of callOpenTiles or getSampleTileMatrix. Both have a path to saved files.}

\item{returnType}{A Boolean. If True, returns object type (OpenTiles, SampleTileMatrix, or NULL). If FALSE, returns True/FALSE}
}
\value{
A string or list of strings in the format 'chr1:100-200' representing
 ranges in the input GRanges
}
\description{
\code{isMOCHAObject} Checks with the object is a MOCHA object, and/or returns the type of object
}
\keyword{utils}
