% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearModeling.R
\name{linearModeling}
\alias{linearModeling}
\title{\code{linearModeling}}
\usage{
linearModeling(
  Obj,
  formula,
  CellType,
  threshold = 0,
  NAtoZero = FALSE,
  numCores = 1
)
}
\arguments{
\item{Obj}{A RangedSummarizedExperment generated from getSampleTileMatrix}

\item{formula}{Formula used for linear modeling.}

\item{CellType}{The name of the celltype you wish to model. Should align with assayNames of the Obj.}

\item{threshold}{A number greater than 0 and less then or equal to 1. The threshold used to determine whether to model a region or not, based on the fraction of non-zero measurements across samples at that location.}

\item{NAtoZero}{Boolean, whether to convert NA region (no accessibility measurement) to zero.}

\item{numCores}{Number of threads to parallelize modeling over. Default is 1.}
}
\value{
A list of lmer model objects
}
\description{
\code{linearModeling} Runs linearModeling on MOCHA TSAM
}
\keyword{downstream}
