% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prediction.R
\name{make_prediction}
\alias{make_prediction}
\title{\code{determine_dynamic_range}}
\usage{
make_prediction(X, finalModelObject)
}
\arguments{
\item{X}{an intensityMatrix output from \code{calculate_intensities}}

\item{finalModel}{is a matrix with coefficients and an index indicating
the number of cell used to train that model}
}
\value{
the original intensityMatrix with the two intensity parameters required
to calculate the probability of a (+) peak, with an additional two columns
that include the prediction probability
}
\description{
\code{make_prediction} is an R helper function, part of the single-cell peak calling
algorithm scMACS by (XXX et al, 2021) that determines which genomic regions, or bins,
will be used for de-novo peak calling. The function "make_prediction" applies the
logistic regression model predictions
}
\details{
The technical details of the algorithm are found in XX.
}
\references{
XX
}
