% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_lambda2.R
\name{max_nbinom_mass_function}
\alias{max_nbinom_mass_function}
\title{\code{Helper Functions}}
\usage{
max_nbinom_mass_function(mean.param, dispersion.param, n, k)
}
\description{
\code{Helper Functions} is a set of helper functions used to determine maximum 
             intensity transformations into an estimate of lambda. These helper functions 
             are called internally and not meant to be called directly outside of the 
             calculate_intensities function call.
}
\details{
The technical details of the algorithm are found in XX.
}
\references{
XX
}
