% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{MochaTileResults-class}
\alias{MochaTileResults-class}
\alias{MochaTileResults}
\title{MOCHA S4 subclasses for tileResults and SampleTileMatrix objects.}
\description{
Subclass of \code{\link[MultiAssayExperiment]{MultiAssayExperiment}} returned
optionally from \code{\link{callOpenTiles}} when \code{returnClass = "mocha"}.
}
