% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromVAR_wrappers.R
\name{modelDeviations}
\alias{modelDeviations}
\title{\code{modelDeviations}}
\usage{
modelDeviations(Obj, formula, type = "z", numCores = 1)
}
\arguments{
\item{Obj}{chromVAR output, from a MOCHA object. Should come from runChromVAR}

\item{formula}{formula for the linear model}

\item{type}{Type of chromVAR output to model. Should be either 'deviations' or 'z' score.}

\item{numCores}{}

\item{verbose}{Set TRUE to display additional messages. Default is FALSE.}
}
\value{
lmem_res A list of linear model results for each chromVAR deviations.
}
\description{
\code{modelDeviations} Runs a generalized linear model on ChromVAR deviation outputs.
}
\details{
relies on lmerTest to run a linear model on each chromVAR result.
}
