% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotifEnrichment.R
\name{MotifEnrichment}
\alias{MotifEnrichment}
\title{\code{MotifEnrichment}}
\usage{
MotifEnrichment(Group1, Group2, motifPosList, type = NULL)
}
\arguments{
\item{Group1}{A GRanges object, such as a set of significant differential
tiles.}

\item{Group2}{A GRanges object containing background regions, non-overlapping
with Group1}

\item{motifPosList}{A GRangesList of motifs and positions for each motif.
Must be named for each motif.}

\item{type}{Optional, name of a metadata column in Group1 and Group2 to test
for enrichment the number of unique entries in column given by 'type'.
Default is NULL, which tests the number of Ranges.}
}
\value{
A data.frame containing enrichment for each group
}
\description{
Test for enrichment of motifs within Group1 against a background
  Group2 using a hypergeometric t-test.
}
