% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mocha_subclass_methods.R
\name{openTiles}
\alias{openTiles}
\title{Get open tiles from a MochaTileResults object}
\usage{
openTiles(
  x,
  cellPopulations = "all",
  returnType = c("GRangesList", "data.frame")
)
}
\arguments{
\item{x}{A \code{MochaTileResults} object.}

\item{cellPopulations}{Cell populations to return, or \code{"all"}.}

\item{returnType}{\code{"GRangesList"} or \code{"data.frame"}.}
}
\value{
Open tiles per cell population.
}
\description{
Get open tiles from a MochaTileResults object
}
\keyword{utils}
