% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bipartite_graph.R
\name{plot_bipartite_graph}
\alias{plot_bipartite_graph}
\title{\code{plot_bipartite_graph}}
\usage{
plot_bipartite_graph(g, circleLayout = T, directed = F, tname = NULL)
}
\arguments{
\item{g:}{data.frame with 2 columns, 'From' = starting vertex; 'To' = ending vertex}

\item{circleLayout:}{boolean indicating whether to use a circle or bipartite layout}

\item{directed:}{boolean indicating whether to}

\item{tname:}{title name}
}
\description{
\code{plot_bipartite_graph} allows you to plot to networks 
             as a bipartite graph
}
