% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_coaccessibility.R
\name{plot_coaccessibility}
\alias{plot_coaccessibility}
\title{\code{plot_coaccessibility}}
\usage{
plot_coaccessibility(zi_mat, index, peak_mat, fname, addLabel = F)
}
\arguments{
\item{mat:}{sample-peak matrix with regions to analyze}

\item{numCores:}{integer to determine # of paralell cores}
}
\value{
a 3-column data.frame containing 
        - zi_mat= Melted Zero-inflated Spearman Correlation Matrix
        - index = numeric integer indicating which combination to plot
        - peak_mat= peak-sample matrix
}
\description{
\code{plot_coaccessibility} allows you to visually determine 
             whether two regions are correlated.
}
\details{
The technical details of the zero-inflated correlation can be
         found here: 

              Pimentel, Ronald Silva, "Kendall's Tau and Spearman's Rho 
              for Zero-Inflated Data" (2009). Dissertations. 

         while the implementation (scHOT R package), can be found here:
              http://www.bioconductor.org/packages/release/bioc/html/scHOT.html
}
\references{
Pimentel, Ronald Silva, "Kendall's Tau and Spearman's Rho 
              for Zero-Inflated Data" (2009). Dissertations.
}
