% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_differntial_region.R
\name{plot_differential_region}
\alias{plot_differential_region}
\title{\code{plot_differential_region}}
\usage{
plot_differential_region(tmp_mat, tileID, group)
}
\arguments{
\item{tmp_mat:}{sample-peak matrix}

\item{tileID:}{chromosome region to analyze}

\item{group:}{vector of group indices (0,1)


The output is a ggplot2 object showing the histograms
of the intensities at a region, by each condition.}
}
\description{
\code{plot_differential_region} allows you to graph a particular
             region across two different conditions
}
\references{
XX
}
