% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Region.R
\name{plot_geneBody}
\alias{plot_geneBody}
\title{Plot a genebody from a specific database}
\usage{
plot_geneBody(
  organismdb,
  geneBody_gr,
  x_lim = NULL,
  base_size = 12,
  theme_ls = .gene_plot_theme,
  collapseGenes = FALSE
)
}
\arguments{
\item{organismdb}{Database object, for example the output of `OrganismDbi::makeOrganismDbFromTxDb()`}

\item{geneBody_gr}{A GRanges object containing the gene, for example an output of `get_grange_genebody()`}

\item{base_size}{Numeric, default 12. Global plot base text size parameter}

\item{collapseGenes}{Logical value, default FALSE. If TRUE will collapse all genes into one line.
sites into a single row}

\item{plot_theme}{Named list of `ggplot2::theme()` parameters.}
}
\value{
A ggplot object
}
\description{
Used in `plotRegion()` to plot gene track for granges input containing all genes in a region
}
