% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_Region.R
\name{plot_pal}
\alias{plot_pal}
\title{Plot colors in a color palette}
\usage{
plot_pal(pal, cex = 4, cex.lab = 1)
}
\arguments{
\item{pal}{A palette, named list of values where names are colors and values are the color breaks}
}
\value{
Plots a base plot to visualize the input palette
}
\description{
Utility function to quickly visualize a color palette formatted for motif weight visualization
}
\examples{
motif_color_pal <- c(-20, -10, 0, 10, 20)
names(motif_color_pal) <- c("darkblue", "cyan", "gray", "orange", "darkred")
plot_pal(motif_color_pal)
}
