% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sample_specific_reproducible.R
\name{plot_sample_specific_reproducible_peaks.R}
\alias{plot_sample_specific_reproducible_peaks.R}
\title{\code{plot_sample_specific_reproducible_peaks.R}}
\usage{
plot_sample_specific_reproducible_peaks.R(
  peak_mat,
  save_plot = FALSE,
  fname = "tmp.png"
)
}
\arguments{
\item{peak_mat}{a peak-sample matrix}

\item{fname}{a string containing where the plots wills be saved
the samples}

\item{index}{a vector of 0, 1, denoting the two groups}
}
\value{
a plot indicating # of common peaks by group using 
        proportion-based thresholding
}
\description{
\code{plot_sample_specific_reproducible_peaks.R} allows you to
             visualize common peaks by setting a proportion-based threshold
}
