% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bipartite_graph.R
\name{plot_tripartite_graph}
\alias{plot_tripartite_graph}
\title{\code{plot_tripartite_graph}}
\usage{
plot_tripartite_graph(df, directed = T, tname = "Ligand-Motif-Gene")
}
\arguments{
\item{g:}{data.frame with 2 columns, 'From' = starting vertex; 'To' = ending vertex}

\item{directed:}{boolean indicating whether to}

\item{tname:}{title name}
}
\description{
\code{plot_bipartite_graph} allows you to plot to networks 
             as a bipartite graph
}
