% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingUtils.R
\name{plot_whichGene}
\alias{plot_whichGene}
\title{Generate a ggplot based on a Gene Model}
\usage{
plot_whichGene(
  newModel,
  x_lim = NULL,
  base_size = 12,
  theme_ls = .gene_plot_theme
)
}
\arguments{
\item{base_size}{Numeric, default 12. Global plot base text size parameter}

\item{theme_ls}{Named list of parameters passed to `theme()`. For defaults see `.gene_plot_theme`}

\item{A}{gene granges model, ie generated by `get_gene_body_model()`}

\item{x_min}{Numeric value. Min x-axis limit. If NULL (default) will use min value from newModel granges.}

\item{x_max}{Numeric value. Max x-axis limit. If NULL (default) will use max value from newModel granges.}
}
\description{
Creates a gene track ggplot based on gene model GRanges
Used conditinally in `plotRegion()` for gene plot.
}
