% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assessDropout.R
\name{plotDropoutDiagnostics}
\alias{plotDropoutDiagnostics}
\title{Plot dropout model diagnostics}
\usage{
plotDropoutDiagnostics(
  TSAM_Object,
  cellPopulation,
  type = c("calibration", "probHist", "perSample")
)
}
\arguments{
\item{TSAM_Object}{A SummarizedExperiment with dropout results from
\code{\link{assessDropout}}.}

\item{cellPopulation}{Cell population to plot.}

\item{type}{One of \code{"calibration"}, \code{"probHist"}, or \code{"perSample"}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot dropout model diagnostics
}
\examples{
\dontrun{
MOCHA::plotDropoutDiagnostics(TSAM_with_dropout, "CD16 Mono", type = "calibration")
}

}
\keyword{plotting}
