% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeFootprints.R
\name{plotMotifs}
\alias{plotMotifs}
\title{Generate motif footprint metrics and/or statistics across samples, cell types, and motifs}
\usage{
plotMotifs(
  motifSE,
  footprint = NULL,
  groupColumn = NULL,
  returnDF = FALSE,
  returnPlotList = FALSE,
  plotIndividualRegions = TRUE,
  relHeights = c(0.3, 0.7)
)
}
\arguments{
\item{motifSE}{A SummarizedExperiment with motif footprinting information, from motifFootprint}

\item{footprint}{Optional string, to describe which footprint within the motifSE to analyze. Default is NULL, at which point it will pull from all footprint present.}

\item{groupColumn}{An optional string, that will contain the group-level labels for either calling footprints as significant, or for comparing motif footprintings stats betweem groups. Default is null, at which point no wilcoxon tests will be conducted. groupColumn can be a sample-level group from the MOCHA's object colData slot, 'CellType' (if you want to compare across cell types), or 'Motif' if you want to compare across motifs.}

\item{callFootprint}{A boolean, to determine whether to call footprints as significant across samples. Requires sample-specific motif footprinting.}

\item{motifSize}{A number describing the approximate motif size you want to use, when looking at the insertion rate over the motif core. Default is 6 (3 bp up and 3 bp downstream)}

\item{searchWidth}{A number, describing the maximum motif footprint width. The function will look for the maximum insertions within +/- this number of the motif center. Default is 50.}

\item{background}{An optional string if you want to compare two different conditions. This string is the background group that you want to use for running wilcoxon tests.}
}
\value{
a data.frame, containing motif footprint stats.
}
\description{
Identifies the width and depth of a given motif for a given group & cell type. Requires the output of motifFootprint. If a groupColumn and background is provided, the function will also conduct wilcoxon tests on the motif baseline, depth, and width between groups.
}
\examples{
\dontrun{
summaryStats <- MOCHA::motifStats(
  motifSE,
  footprint = "CD4 Naive_ARID5A", groupColumn = 'COVID_status', background = 'Negative')
}

}
\keyword{exporting}
