% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetMOCHAObject.R
\name{renameCellTypes}
\alias{renameCellTypes}
\title{\code{renameCellTypes}}
\usage{
renameCellTypes(MOCHAObject, oldNames, newNames)
}
\arguments{
\item{MOCHAObject}{A  RangedSummarizedExperiment,}

\item{oldNames}{A list of cell type names that you want to change.}

\item{newNames}{A list of new cell type names to replace the old names with.}
}
\value{
A MOCHA SampleTile object with new cell types.
}
\description{
\code{renameCellTypes} Allows you to modify the cell type names for a MOCHA SampleTileObject, from
                   the assay names, GRanges column names, and summarizedData (within the metadata), all at once.
}
