% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromVAR_wrappers.R
\name{runChromVAR}
\alias{runChromVAR}
\title{\code{runChromVAR}}
\usage{
runChromVAR(TSAM_Object, motifSetName, cellTypeSpecific = FALSE)
}
\arguments{
\item{TSAM_Object}{RangedSummarizedExperiment object, containing the TSAM from getSampleTileMatrix, with a motif set added via addMotifSet()}

\item{motifSetName}{name of the motifList in the SE_object's metadata. Must be present to run chromVAR}

\item{groupList}{the list of cell type names or sample-associated data that
should be used to subset the Object}

\item{na.rm}{removes groups that are NA if set to true. If set to false, then
you filter for everything in the groupList and also NA values.}
}
\value{
chromVAR object
}
\description{
\code{runChromVAR} Runs ChromVAR on a MOCHA Tile-Sample Accessibility Matrix object
}
\details{
This is a wrapper for basic SummarizedExperiment-based subsetting.
}
