% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlePopulationConsensusTiles.R
\name{simplifiedConsensusTiles}
\alias{simplifiedConsensusTiles}
\title{\code{simplifiedConsensusTiles}}
\usage{
simplifiedConsensusTiles(ref)
}
\arguments{
\item{ref}{ref a list of data on one specific cell type, sampleData, threshold, groupColumn, and verbose flag.}
}
\description{
\code{simplifiedConsensusTiles} is an R helper function, part of
  the single-cell peak
}
\keyword{internal}
