% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingUtils.R
\name{simplifiedOrgDb}
\alias{simplifiedOrgDb}
\title{Generate a simplified Organism Db object.}
\usage{
simplifiedOrgDb(TxDb = TxDb, orgdb = orgdb)
}
\arguments{
\item{TxDb}{A TxDb object with transcript info.}

\item{orgdb}{A OrgDb object with gene name info.}
}
\description{
Pulls out only the longest transcripts for each gene and repackages them into an organism database object for plotting.
}
