% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlePopulationSampleTileMatrix.R
\name{simplifiedSampleTile}
\alias{simplifiedSampleTile}
\title{\code{singlePopulationSampleTileMatrix}}
\usage{
simplifiedSampleTile(ref)
}
\arguments{
\item{x}{ref a list of data on one specific cell type (RaggedExperiment) and all tiles that were called across that given population.}
}
\value{
sampleTileIntensityMat a sample X peak matrix containing observed
  measurements for each sample at each peak.
}
\description{
\code{singlePopulationSampleTileMatrix} is a function that can transform
  a set of tile intensities into peak X sample matrix for a custom set of tiles
}
\details{
The technical details of the algorithm are found in singlePopulationSampleTileMatrix
}
\references{
XX
}
\keyword{internal}
