% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateFragments.R
\name{simulateFragments}
\alias{simulateFragments}
\title{simulate scATAC-seq fragments}
\usage{
simulateFragments(
  nCells = 500,
  meanFragsPerCell = 5000,
  fragThreshold = 1000,
  allLocationsGR = getSimulatedPeakSet(),
  peakCenters = 1,
  FRIP = 0.9,
  meanLengths = c(75, 200),
  lengthProbability = c(0.9, 0.1)
)
}
\arguments{
\item{nCells}{Number of cells to simulate (i.e. size of the cell population) Default is 500}

\item{meanFragsPerCell}{Mean number of fragments to be assigned per cell. A distribution will be build around this. Default is 5000}

\item{fragThreshold}{Minimum number of fragments per cell. This mimic QC thresholding for poor quality cells. 
Cells with fewer than this number of fragments will be regenerated. Default is 1000.}

\item{allLocationsGR}{A GenomicRanges object, generated by getSimulatedPeakSet. It must include all tiles across the genome, including both open and closed regions.}

\item{peakCenters}{The number of centers to a given peak within the peak windows. This is used to assign fragments to peaks. Default is 1.}

\item{FRIP}{The fraction of reads in peaks. Default is 0.9}

\item{meanLengths}{A vector of mean fragments lengths. Fragments lengths from scATAC-seq usually involve a diminishing period distribution, with
an initial distribution around 75 bp, and small number of peaks distributed around 200 bp, and further. For default, we do c(75,200)}

\item{lengthProbability}{A vector that adds up to 1 and is the same length as meanLengths. 
This vector will describe the relative probability of a fragment belonging to a distribution around}
}
\value{
A GenomicRanges object of all fragments, including settings used to generate them and Cell barcodes.
}
\description{
\code{simulateFragments} Simulated single cell ATAC-seq fragments from a single population of cells.
}
