% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateFragments.R
\name{simulateMetaData}
\alias{simulateMetaData}
\title{simulate scATAC-seq metadata}
\usage{
simulateMetaData(simulatedFragmentData, sampleNumber = 1)
}
\arguments{
\item{simulatedFragmentData}{A GenomicRanges object from simulateFragments}

\item{sampleNumber}{1}
}
\value{
A data.frame of simulated cellular metadata
}
\description{
\code{simulateMetaData} Simulated single cell ATAC-seq metadata, created to matched simulated fragment data
}
