% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlePopulationSampleTileMatrix.R
\name{singlePopulationSampleTileMatrix}
\alias{singlePopulationSampleTileMatrix}
\title{\code{singlePopulationSampleTileMatrix}}
\usage{
singlePopulationSampleTileMatrix(
  peaksExperiment,
  consensusTiles,
  NAtoZero = FALSE
)
}
\arguments{
\item{peaksExperiment}{peakset RaggedExperiment, one celltype from the output of callOpenTiles}

\item{consensusTiles}{a vector containing the tileIDs to subset the
sample-tile matrix}
}
\value{
sampleTileIntensityMat a sample X peak matrix containing observed
  measurements for each sample at each peak.
}
\description{
\code{singlePopulationSampleTileMatrix} is a function that can transform
  a set of tile intensities into peak X sample matrix for a custom set of tiles
}
\details{
The technical details of the algorithm are found in XX.
}
\references{
XX
}
\keyword{internal}
