% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{StringsToGRanges}
\alias{StringsToGRanges}
\title{\code{StringsToGRanges}}
\usage{
StringsToGRanges(regionString)
}
\arguments{
\item{regionString}{A string or list of strings each in the format chr1:100-200}
}
\value{
a GRanges object with ranges representing the input string(s)
}
\description{
\code{StringsToGRanges} Turns a list of strings in the format chr1:100-200
  into a GRanges object
}
