% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromVAR_wrappers.R
\name{subsetDev}
\alias{subsetDev}
\title{\code{subsetDev}}
\usage{
subsetDev(Object, subsetBy, groupList)
}
\arguments{
\item{subsetBy}{the variable to subset by. Can either be 'celltype', or a
column from the sample metadata (see colData(Object))}

\item{groupList}{the list of cell type names or sample-associated data that
should be used to subset the Object}

\item{Obj}{chromVAR output, from a MOCHA object. Should come from runChromVAR}

\item{na.rm}{removes groups that are NA if set to true. If set to false, then
you filter for everything in the groupList and also NA values.}
}
\value{
Object the input Object, filtered down to either the cell type or
  samples desired.
}
\description{
\code{subsetDev} subsets a chromVAR output by a given metadata column (subsetBy) and the specific groups within (groupList)
}
\details{
This is a wrapper for basic SummarizedExperiment-based subsetting.
}
