% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_part.R
\name{TwoPart}
\alias{TwoPart}
\title{\code{TwoPart}}
\usage{
TwoPart(data, group, test = "t.test", point.mass = 0)
}
\arguments{
\item{data}{is a vector containing the values for the 2-part test}

\item{group}{is a vector containing the two groups for comparison (in binary, 0,1)}
}
\value{
ans is a list containing the test statistic & p-value corresponding to the 
        two-part wilcoxon test.
}
\description{
\code{TwoPart} is the main peak-calling function in scMACS
             that serves as a wrapper function to call peaks provided a set 
             of fragment files and an ArchR Project for meta-data purposes
}
\details{
The code from the two part test is taken from Taylor & pollard (2009) (reference below) which illustrates how to create a composite test from zero-inflated data.
}
\references{
Taylor, Sandra, and Katherine Pollard. "Hypothesis tests for point-mass mixture data #' with application toomics data with many zero values." Statistical applications in genetics and #' molecular biology 8.1 (2009).
}
