% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateDirectoryPath}
\alias{updateDirectoryPath}
\title{updateDirectoryPath}
\usage{
updateDirectoryPath(Object, directoryPath = NULL)
}
\arguments{
\item{Object}{The output of callOpenTiles or getSampleTileMatrix. Both have a path to saved files.}

\item{directoryPath}{A string, containing the absolute path to the directory.}
}
\value{
A string or list of strings in the format 'chr1:100-200' representing
 ranges in the input GRanges
}
\description{
\code{updateDirectoryPath} Updated the path to save files, in case the directory moved.
}
\keyword{utils}
