% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zi_spearman.R
\name{weightedSpearman}
\alias{weightedSpearman}
\title{scHOT functions, supersceded by used of wCorr's weighted correlation function, which backends to C for faster calculations.}
\usage{
weightedSpearman(x, y, w = 1)
}
\description{
scHOT functions, supersceded by used of wCorr's weighted correlation function, which backends to C for faster calculations.
}
