% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zi_spearman.R
\name{weightedZISpearman}
\alias{weightedZISpearman}
\title{weightedZISpearman}
\usage{
weightedZISpearman(x, y, w = 1)
}
\arguments{
\item{x}{x and y are non-negative data vectors}

\item{y}{x and y are non-negative data vectors}

\item{w}{weight vector, values should be between 0 and 1}
}
\value{
\code{numeric} weighted rho* association value between x and y


scHOT 
Ghazanfar, S., Lin, Y., Su, X. et al. Investigating higher-order interactions in 
  single-cell data with scHOT. Nat Methods 17, 799–806 (2020). 
  https://doi.org/10.1038/s41592-020-0885-x

Zero-Inflated Correlation
Pimentel, Ronald Silva, "Kendall's Tau and Spearman's Rho for
  Zero-Inflated Data" (2009). Dissertations. 721.
  https://scholarworks.wmich.edu/dissertations/721
}
\description{
scHOT implemented Pimentel's zero-inflated (ZI) correlation
in their R package, providing implementations 
of the ZI spearman and tau rank correlations. 
scMACS implements a slight modification of 
scHOT's zero-inflated correlation measure, by
returning NAs in the cases where the correlation
is undefined, and modifying it to use a C-backed correlation program. 
Both references are provided below and are
referenced in all documentations to indicate
their work in implementing these methods.
}
\details{
Description: 
the weightedZISpearman function calculates weighted rho\*,
where rho\* is described in Pimentel et al (2009).
This association measure is defined for zero-inflated,
non-negative random variables.
}
