% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{youden_threshold}
\alias{youden_threshold}
\title{youden_threshold}
\format{
A list of 18 regression variables
}
\usage{
youden_threshold
}
\description{
Trained regression model for predicting a cutoff threshold
for peak calling.
Call:
loess(formula = OptimalCutpoint ~ Ncells, data = thresh_df)
}
\details{
Number of Observations: 27
Equivalent Number of Parameters: 5.98
Residual Standard Error: 0.02121
}
\keyword{datasets}
