% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingUtils.R
\name{breaks_to_scaledbreaks}
\alias{breaks_to_scaledbreaks}
\title{Get scaled values for custom breaks}
\usage{
breaks_to_scaledbreaks(breaks, x)
}
\arguments{
\item{x}{vector of weights}

\item{vector}{of breaks}
}
\description{
Get scaled values for custom breaks in a given value vector for use
defining breaks in scale_gradientn(), for example
}
