% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callOpenTiles.R
\name{callOpenTiles}
\alias{callOpenTiles}
\title{\code{callOpenTiles}}
\usage{
callOpenTiles(
  ArchRProj,
  cellPopLabel,
  cellPopulations = "ALL",
  sampleLabel,
  numCores = 30
)
}
\arguments{
\item{ArchRProj}{an ArchR Project}

\item{cellPopLabel}{string indicating which column in the ArchRProject
metadata contains the cell population label.}

\item{cellPopulations}{vector of strings. Cell subsets for which to call
peaks. This list of group names must be identical to names that appear in
the ArchRProject metadata.  Optional, if cellPopulations='ALL', then peak
calling is done on all cell populations in the ArchR project metadata.
Default is 'ALL'.}

\item{sampleLabel}{string indicating which column in the ArchRProject
metadata contains the sample IDs.}

\item{numCores}{integer. Number of cores to parallelize peak-calling across
multiple cell populations}
}
\value{

}
\description{
\code{callOpenTiles} is the main peak-calling function in scMACS
  that serves as a wrapper function to call peaks provided a set of fragment
  files and an ArchR Project for meta-data purposes
}
\details{
The technical details of the algorithm are found in XX.
}
\references{
XX
}
