% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeaks_population.R
\name{callPeaks_by_population}
\alias{callPeaks_by_population}
\title{\code{callPeaks}}
\usage{
callPeaks_by_population(
  ArchRProj,
  cellSubsets = NULL,
  cellCol_label_name = NULL,
  returnAllPeaks = FALSE,
  numCores = 10,
  totalFrags,
<<<<<<< Updated upstream
  fragsList = NULL,
  StudypreFactor = NULL
=======
  fragsList = NULL
>>>>>>> Stashed changes
)
}
\arguments{
\item{ArchRProj}{an ArchR Project}

\item{cellSubsets}{vector of strings. Cell subsets for which to call peaks. Optional, if cellSubsets='ALL', then peak calling is done on all cell populations in the ArchR project metadata}

\item{cellCol_label_name}{string indicating which column in the meta data file contains 
the cell population label}

\item{returnAllPeaks}{boolean. Indicates whether scMACS should return object containing all genomic regions or just the positive (+) called peaks. Default to the latter, only positive peaks.}

\item{numCores}{integer. Number of cores to parallelize peak-calling across
multiple cell populations}

\item{totalFrags}{# of fragments in that sample for that cell population}
<<<<<<< Updated upstream

\item{StudypreFactor}{ratio of average signal between new study and training set}
=======
>>>>>>> Stashed changes
}
\value{
scMACs_PeakList an list containing peak calls for each cell population passed on in the 
        cell subsets argument. Each peak call is returned as as Genomic Ranges object.
}
\description{
\code{callPeaks} is the main peak-calling function in scMACS
             that serves as a wrapper function to call peaks provided a set 
             of fragment files and an ArchR Project for meta-data purposes
}
\details{
The technical details of the algorithm are found in XX.
}
\references{
XX
}
