% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callPeaks_sample.R
\name{callPeaks_by_sample}
\alias{callPeaks_by_sample}
\title{\code{callPeaks_by_sample}}
\usage{
callPeaks_by_sample(
  ArchRProj,
  cellSubsets = NULL,
  cellCol_label_name = NULL,
  sampleCol_label_name = NULL,
  returnAllPeaks = FALSE,
  numCores = 10
)
}
\arguments{
\item{ArchRProj}{an ArchR Project}

\item{cellSubsets}{vector of strings. Cell subsets for which to call peaks. Optional, if cellSubsets='ALL', then peak calling is done on all cell populations in the ArchR project metadata}

\item{cellCol_label_name}{string indicating which column in the meta data file contains 
the cell population label}

\item{returnAllPeaks}{boolean. Indicates whether scMACS should return object containing all genomic regions or just the positive (+) called peaks. Default to the latter, only positive peaks.}

\item{numCores}{integer. Number of cores to parallelize peak-calling across
multiple cell populationsdevto}
}
\value{
scMACs_PeakList an list containing peak calls for each cell population passed on in the 
        cell subsets argument. Each peak call is returned as as Genomic Ranges object.
}
\description{
\code{callPeaks_by_sample} is the main peak-calling function in scMACS
             that serves as a wrapper function to call peaks provided a set 
             of fragment files and an ArchR Project for meta-data purposes
}
\details{
The technical details of the algorithm are found in XX.
}
\references{
XX
}
