% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/co_accessibility.R
\name{co_accessibility}
\alias{co_accessibility}
\title{\code{co_accessibility}}
\usage{
co_accessibility(
  subMat,
  filterPairs = NULL,
  index = NULL,
  numCores = 40,
  verbose = FALSE
)
}
\arguments{
\item{mat:}{sample-peak matrix with regions to analyze}

\item{numCores:}{integer to determine # of paralell cores}
}
\value{
a 3-column data.frame containing 
        - Correlation= Zero-inflated Spearman Correlation
        - Peak1= location of co-accessible region 1
        - Peak2= location of co-accessible region 2
}
\description{
\code{co_accessibility} allows you to determine whether 2 peaks
             are co-accessible using a zero-inflated spearman correlation.
}
\details{
The technical details of the zero-inflated correlation can be
         found here: 

              Pimentel, Ronald Silva, "Kendall's Tau and Spearman's Rho 
              for Zero-Inflated Data" (2009). Dissertations. 

         while the implementation (scHOT R package), can be found here:
              http://www.bioconductor.org/packages/release/bioc/html/scHOT.html
}
\examples{
Generate 
mat1 = matrix(pmax(0, rnorm(1000)), ncol=100); row.names(mat1) <- paste('A',1:10,sep='_')
ziSpear_mat <- co_accessibility(mat1, numCores=5)
head(ziSpear_mat)
}
\references{
XX
}
