% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testCoAccessibility.R
\name{combineSampleTileMatrix}
\alias{combineSampleTileMatrix}
\title{\code{combineSampleTileMatrix}}
\usage{
combineSampleTileMatrix(STObj, NAtoZero = TRUE, verbose = FALSE)
}
\arguments{
\item{STObj}{The SummarizedExperiment object output from getSampleTileMatrix containing your sample-tile matrices}

\item{verbose}{Set TRUE to display additional messages. Default is FALSE.}

\item{NAToZero}{Set NA values in the sample-tile matrix to zero}
}
\value{
TileCorr A data.table correlation matrix
}
\description{
\code{combineSampleTileMatrix} combines all celltypes in a SampleTileMatrix object into a SummarizedExperiment with one single matrix across all cell types and samples.
}
