% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{differentialsToGRanges}
\alias{differentialsToGRanges}
\title{Convert a data.frame or matrix to a GRanges}
\usage{
differentialsToGRanges(differentials, tileColumn = "Tile")
}
\arguments{
\item{differentials}{a matrix/data.frame with a column tileColumn containing
region strings in the format "chr:start-end"}

\item{tileColumn}{name of column containing region strings. Default is "Tile".}
}
\value{
a GRanges containing all original information
}
\description{
Convert a data.frame or matrix to a GRanges
}
\keyword{utils}
