% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCoAccessibleLinks.R
\name{FilterCoAccessibleLinks}
\alias{FilterCoAccessibleLinks}
\title{\code{FilterCoAccessibleLinks}}
\usage{
FilterCoAccessibleLinks(PeakCorr, threshold = 0.5)
}
\arguments{
\item{PeakCorr}{}

\item{threshold}{}
}
\value{

}
\description{
\code{FilterCoAccessibleLinks} does stuff
}
\details{
The technical details of the zero-inflated correlation can be
         found here:

              Pimentel, Ronald Silva, "Kendall's Tau and Spearman's Rho
              for Zero-Inflated Data" (2009). Dissertations.

         while the implementation (scHOT R package), can be found here:
              http://www.bioconductor.org/packages/release/bioc/html/scHOT.html
}
