% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingUtils.R
\name{get_grange_genebody}
\alias{get_grange_genebody}
\title{Get GeneBody GRanges for a given region}
\usage{
get_grange_genebody(regionGRanges, TxDb, single.strand.genes.only = TRUE)
}
\arguments{
\item{regionGRanges}{regionGRanges A region Granges object to retrieve gene bodies for. For example, the output of countdf_to_region.}

\item{TxDb}{A TxDb database}

\item{single.strand.genes.only}{Logical, default FALSE.}
}
\description{
Used in `plotRegion()`
}
