% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoAccessibleLinks.R
\name{getCoAccessibleLinks}
\alias{getCoAccessibleLinks}
\title{\code{getCoAccessibleLinks}}
\usage{
getCoAccessibleLinks(
  peakDT,
  regions,
  windowSize = 1 * 10^6,
  numCores = 1,
  verbose = FALSE
)
}
\arguments{
\item{regions}{a GRanges containing the regions to compute co-accessible links for. 
Can be the output from getDifferentialAccessibleTiles.}

\item{numCores}{Optional, the number of cores to use with multiprocessing. Default is 1.}

\item{SampleTileObj}{The SummarizedExperiment object output from getSampleTileMatrix}

\item{cellPopulation}{A string denoting the cell population of interest}
}
\description{
\code{getCoAccessibleLinks} does stuff
}
\details{
The technical details of the zero-inflated correlation can be
         found here:

              Pimentel, Ronald Silva, "Kendall's Tau and Spearman's Rho
              for Zero-Inflated Data" (2009). Dissertations.

         while the implementation (scHOT R package), can be found here:
              http://www.bioconductor.org/packages/release/bioc/html/scHOT.html
}
