% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifferentialAccessibleTiles.R
\name{getDifferentialAccessibleTiles}
\alias{getDifferentialAccessibleTiles}
\title{Conduct a differential test between open regions of two sample groups}
\usage{
getDifferentialAccessibleTiles(
  SampleTileObj,
  cellPopulation,
  groupColumn,
  foreground,
  background,
  minZeroDiff = 0.5,
  qValueMethod = "experimental",
  signalThreshold = NULL,
  qValueThreshold = 0.2,
  outputGRanges = TRUE,
  numCores = 1,
  verbose = FALSE
)
}
\arguments{
\item{SampleTileObj}{The SummarizedExperiment object output from
getSampleTileMatrix}

\item{cellPopulation}{A string denoting the cell population of interest}

\item{groupColumn}{The column containing sample group labels}

\item{foreground}{The foreground group of samples for differential comparison}

\item{background}{The background group of samples for differential comparison}

\item{minZeroDiff}{Minimum difference in average dropout rates across groups
require to keep tiles for differential testing. Default is 0.5 (50\%).}

\item{qValueMethod}{String describing qvalue method. Can be 'standard', or 'experimental'. 
See methods in the MOCHA manuscript for description of the experimental. 
Otherwise, 'standard' applies standard q value.}

\item{signalThreshold}{Optional value. 
This is Minimum median intensity required to keep tiles for
differential testing to increase statistical power in small sample cohorts.
Default is NULL, at which point the optimal threshold will be found.}

\item{qValueThreshold}{A number greater than 0 and less than 1, used to optimize the noise thresholding parameter by cell type. Default is 0.2. It is recommended to keep this paramter between 0.05 and 0.3.}

\item{outputGRanges}{Outputs a GRanges if TRUE and a data.frame if
FALSE. Default is TRUE.}

\item{numCores}{The number of cores to use with multiprocessing.
Default is 1.}

\item{verbose}{Set TRUE to display additional messages. Default is FALSE.}
}
\value{
full_results The differential accessibility results as a GRanges or
  matrix data.frame depending on the flag `outputGRanges`.
}
\description{
\code{getDifferentialAccessibleTiles} allows you to
  determine whether regions of chromatin are differentially accessible
  between groups by conducting a test
}
\examples{
\dontrun{
cellPopulation <- "MAIT"
foreground <- "Positive"
background <- "Negative"
# Choose to output a GRanges or data.frame.
# Default is TRUE
outputGRanges <- TRUE
# SampleTileMatrices is the output of MOCHA::getSampleTileMatrix
differentials <- MOCHA::getDifferentialAccessibleTiles(
  SampleTileObj = SampleTileMatrices,
  cellPopulation = cellPopulation,
  groupColumn = groupColumn,
  foreground = foreground,
  background = background,
  outputGRanges = outputGRanges,
  numCores = numCores
)
}
}
\keyword{downstream}
