% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifferentialAccessibleTiles.R
\name{getDifferentialAccessibleTiles}
\alias{getDifferentialAccessibleTiles}
\title{\code{getDifferentialAccessibleTiles}}
\usage{
getDifferentialAccessibleTiles(
  SampleTileObj,
  cellPopulation,
  groupColumn,
  foreground,
  background,
  fdrToDisplay = 0.2,
  numCores = 2
)
}
\arguments{
\item{SampleTileObj}{The SummarizedExperiment object output from getSampleTileMatrix}

\item{cellPopulation}{A string denoting the cell population of interest}

\item{groupColumn}{The column containing sample group labels}

\item{foreground}{The foreground group of samples for differential comparison}

\item{background}{The background group of samples for differential comparison}

\item{fdrToDisplay}{Optional, false-discovery rate used only for standard output messaging}

\item{numCores}{Optional, the number of cores to use with multiprocessing. Default is 1.}
}
\description{
\code{getDifferentialAccessibleTiles} allows you to
  determine whether regions of chromatin are differentially accessible
  between groups by conducting a test
}
