% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifferentialAccessibleTiles.R
\name{getDifferentialAccessibleTiles}
\alias{getDifferentialAccessibleTiles}
\title{\code{get_differential_accessible_Regions}}
\usage{
getDifferentialAccessibleTiles(
  sampleTileMatrix,
  tileResults,
  fdr_control = 0.2,
  numCores = 2
)
}
\description{
\code{get_differential_accessible_Regions} allows you to
  determine whether regions of chromatin are differentially accessible
  between groups by conducting a test
}
