% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDifferentialAccessibleTiles.R
\name{getDifferentialAccessibleTiles}
\alias{getDifferentialAccessibleTiles}
\title{\code{getDifferentialAccessibleTiles}}
\usage{
getDifferentialAccessibleTiles(
  SampleTileObj,
  cellPopulation,
  groupColumn,
  foreground,
  background,
  fdrToDisplay = 0.2,
  outputGRanges = TRUE,
  numCores = 2
)
}
\arguments{
\item{SampleTileObj}{The SummarizedExperiment object output from getSampleTileMatrix}

\item{cellPopulation}{A string denoting the cell population of interest}

\item{groupColumn}{The column containing sample group labels}

\item{foreground}{The foreground group of samples for differential comparison}

\item{background}{The background group of samples for differential comparison}

\item{fdrToDisplay}{Optional, false-discovery rate used only for standard output messaging}

\item{outputGRanges}{Optional, outputs a GRanges if TRUE and a data.frame if FALSE.
Default is TRUE.}

\item{numCores}{Optional, the number of cores to use with multiprocessing. Default is 1.}
}
\value{
full_results The differential accessibility results as a GRanges or matrix
  data.frame depending on the flag `outputGRanges`.
}
\description{
\code{getDifferentialAccessibleTiles} allows you to
  determine whether regions of chromatin are differentially accessible
  between groups by conducting a test
}
\examples{
\dontrun{
regions <- head(differentials, 10)

# Alternatively, define regions as a character vector
# of region strings in the format "chr:start-end"
regions <- c(
  "chrY:7326500-7326999",
  "chrY:7327000-7327499",
  "chrY:7339500-7339999",
  "chrY:7344500-7344999"
)
links <- scMACS::getCoAccessibleLinks(
  SampleTileObj = SampleTileMatricesAnnotated,
  cellPopulation = cellPopulation,
  regions = regions,
  windowSize = 1 * 10^6,
  numCores = numCores,
  verbose = TRUE
)
}
}
