% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runZIGLMM.R
\name{getModelValues}
\alias{getModelValues}
\title{getModelValues from runZIGLMM output.}
\usage{
getModelValues(object, specificVariable)
}
\arguments{
\item{object}{A SummarizedExperiment object generated from runZIGLMM.}

\item{specificVariable}{A string, describing the factor of influence.}
}
\value{
A data.frame of slopes, significance, and standard error for one factor.
}
\description{
\code{getModelValues} Pull out a data.frame of model values (slope, significance, and std.error) for a given factor from the SummarizedExperiment output of runZIGLMM.
}
\examples{
\dontrun{
age_df <- getModelValues(runZIGLMM_output, "Age")
}

}
