% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSampleTileMatrix.R
\name{getSampleTileMatrix}
\alias{getSampleTileMatrix}
\title{\code{create_peak_sampleMatrix}}
\usage{
getSampleTileMatrix(
  tileResults,
  cellPopulations = "ALL",
  groupColumn = NULL,
  threshold = 0.2,
  join = "union",
  NAtoZero = FALSE,
  log2Intensity = TRUE
)
}
\arguments{
\item{tileResults}{output of callOpenTiles}
}
\value{
sampleTileIntensityMat a sample X peak matrix containing observed
  measurements for each sample at each peak.
}
\description{
\code{create_peak_sampleMatrix} is a function that can transform
  a set of sample-specific peak calls into peak X sample matrix containing
  lambda1 measurements for each sample.
}
\details{
The technical details of the algorithm are found in XX.
}
\references{
XX
}
