% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GRangesToString}
\alias{GRangesToString}
\title{\code{GRangesToString}}
\usage{
GRangesToString(GR_obj)
}
\description{
\code{GRangesToStrin} Turns a GRanges Object into
	a list of strings in the format chr1:100-200
}
