% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearModeling.R
\name{linearModeling}
\alias{linearModeling}
\title{\code{linearModeling}}
\usage{
linearModeling(
  Obj,
  formula,
  CellType,
  threshold = 0,
  NAtoZero = FALSE,
  numCores = 1
)
}
\arguments{
\item{Obj}{A RangedSummarizedExperment generated from getSampleTileMatrix}

\item{Donor_col}{The metadata column with donor information.}

\item{Time_col}{The metadata column with longitudinal information}

\item{Group_col}{The metadata column with group information. Default is NULL, in which case it will attempt to use celltype as a proxy. Do not run group and cell type at the same time.}

\item{returnObj}{Boolean flag for whether or not you want the full PALMO object. Default is FALSE, in which case it returns the variance decomposition data.frame.}
}
\value{
variance decomposition matrix
}
\description{
\code{linearModeling} Runs linearModeling on MOCHA TSAM
}
\examples{
\dontrun{
     vd_mat <- varDecomp(SampleTileObj, Donor_col = 'PTID', Time_col = 'Days')
)
}

}
