% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plotIntensityDistribution}
\alias{plotIntensityDistribution}
\title{\code{plotIntensityDistribution}}
\usage{
plotIntensityDistribution(TSAM_object, cellPopulation, returnDF = FALSE)
}
\arguments{
\item{TSAM_object}{SummarizedExperiment from getSampleTileMatrix}

\item{cellPopulation}{Cell type names (assay name) within the TSAM_object}
}
\value{
data.frame or ggplot histogram.
}
\description{
\code{plotIntensityDistribution}  Plots the distribution of sample-tile intensities for a give cell type
}
