% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runZIGLMM.R
\name{runZIGLMM}
\alias{runZIGLMM}
\title{Run Zero-Inflated Generalized Linear Mixed-Modeling for zero inflated
  data}
\usage{
runZIGLMM(
  TSAM_Object,
  cellPopulation = NULL,
  continuousFormula = NULL,
  ziformula = NULL,
  initialSampling = 5,
  verbose = FALSE,
  numCores = 1
)
}
\arguments{
\item{TSAM_Object}{A SummarizedExperiment object generated from
getSampleTileMatrix, chromVAR, or other. It is expected to contain only one
assay, or only the first assay will be used for the model. Data should not
be zero-inflated.}

\item{cellPopulation}{A single cell population on which to run this model}

\item{continuousFormula}{The formula, see \code{\link[glmmTMB]{glmmTMB}}.
Combined fixed and random effects formula, following lme4 syntax.}

\item{ziformula}{The zero-inflated formula, see
\code{\link[glmmTMB]{glmmTMB}}. a one-sided (i.e., no response variable)
formula for zero-inflation combining fixed and random effects: the default
~0 specifies no zero-inflation. Specifying ~. sets the zero-inflation
formula identical to the right-hand side of formula (i.e., the conditional
effects formula); terms can also be added or subtracted. When using ~. as
the zero-inflation formula in models where the conditional effects formula
contains an offset term, the offset term will automatically be dropped. The
zero-inflation model uses a logit link.}

\item{initialSampling}{Size of data to use for pilot}

\item{verbose}{Set TRUE to display additional messages. Default is FALSE.}

\item{numCores}{integer. Number of cores to parallelize across.}
}
\value{
results a SummarizedExperiment containing LMEM results
}
\description{
\code{runZIGLMM} Runs linear mixed-effects modeling for
  continuous, non-zero inflated data using \code{\link[glmmTMB]{glmmTMB}}
}
\examples{
\dontrun{
  modelList <- runLMEM(ExperimentObj,
    modelFormula = NULL,
    initialSampling = 5,
    verbose = FALSE,
    numCores = 1
 )
}

}
\keyword{downstream}
