% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shareMOCHA.R
\name{shareMOCHA}
\alias{shareMOCHA}
\title{\code{packMOCHA}}
\usage{
shareMOCHA(MOCHA_Obj, zipName = "ZippedMOCHA.zip")
}
\arguments{
\item{zipName}{zipName Name of the compressed file that will be exported.}

\item{MOCHA_Object}{A MultiAssayExperiment or RangedSummarizedExperiment, from MOCHA}
}
\value{
path to zipped object.
}
\description{
\code{packMOCHA} compresses a MOCHA sample-tile object or tile results object, along with the coverage tracks necessary for plotting, 
     and exports
}
