% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singlePopulationConsensusTiles.R
\name{singlePopulationConsensusTiles}
\alias{singlePopulationConsensusTiles}
\title{\code{singlePopulationConsensusTiles}}
\usage{
singlePopulationConsensusTiles(
  peaksExperiment,
  cellPopulation,
  threshold,
  groupColumn,
  join
)
}
\description{
\code{singlePopulationConsensusTiles} is an R helper function, part of
  the single-cell peak
}
\keyword{internal}
