% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetMOCHAObject.R
\name{subsetMOCHAObject}
\alias{subsetMOCHAObject}
\title{Subsets a tileResults-type object by metadata}
\usage{
subsetMOCHAObject(
  Object,
  subsetBy,
  groupList,
  na.rm = TRUE,
  subsetPeaks = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{Object}{A MultiAssayExperiment or RangedSummarizedExperiment,}

\item{subsetBy}{the variable to subset by. Can either be 'celltype', or a
column from the sample metadata (see colData(Object))}

\item{groupList}{the list of cell type names or sample-associated data that
should be used to subset the Object}

\item{na.rm}{removes groups that are NA if set to true. If set to false, then
you filter for everything in the groupList and also NA values.}

\item{subsetPeaks}{If subsetting by cell types, then you need to decide if you want to subset the tile set down to tiles 
only called in those cell types. The default is TRUE.}

\item{verbose}{Set TRUE to display additional messages. Default is FALSE.}
}
\value{
Object the input Object, filtered down to either the cell type or
  samples desired.
}
\description{
\code{subsetObject} subsets a tileResults-type object (from
  callOpenTiles), or a SummarizedExperiment-type object (from
  getSampleTileMatrix), either by cell type or sample metadata.
}
\keyword{utils}
