% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetMOCHAObject.R
\name{subsetMOCHAObject}
\alias{subsetMOCHAObject}
\title{\code{subsetObject}}
\usage{
subsetMOCHAObject(Object, subsetBy, groupList, na.rm = TRUE)
}
\arguments{
\item{Object}{A MultiAssayExperiment or RangedSummarizedExperiment,}

\item{subsetBy}{the variable to subset by. Can either be 'celltype', or a
column from the sample metadata (see colData(Object))}

\item{groupList}{the list of cell type names or sample-associated data that
should be used to subset the Object}

\item{na.rm}{removes groups that are NA if set to true. If set to false, then
you filter for everything in the groupList and also NA values.}
}
\value{
Object the input Object, filtered down to either the cell type or
  samples desired.
}
\description{
\code{subsetObject} subsets a tileResults-type object (from
  callOpenTiles), or a SummarizedExperiment-type object (from
  getSampleTileMatrix), either by cell type or sample metadata.
}
