% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shareMOCHA.R
\name{unpackMOCHA}
\alias{unpackMOCHA}
\title{\code{unpackMOCHA}}
\usage{
unpackMOCHA(packedObject = NULL, outDir = getwd())
}
\arguments{
\item{packedObject}{The path to the packed MOCHA object.}

\item{outDir}{The path to the external directory where you want to unpack the MOCHA object.}
}
\value{
the MOCHA object (tile results or Sample tile matrix)
}
\description{
\code{unpackMOCHA} Take a packed MOCHA object, and unpacks it locally, while resetting the directory path to the new location.
}
